Refactor example into three seperate example files
